Checkpoint of code review fixes
