+++ packages/cli/src/ui/hooks/toolMapping.ts
@@ -26,10 +26,12 @@ export function mapToDisplay(
   options: {
     borderTop?: boolean;
     borderBottom?: boolean;
+    borderColor?: string;
+    borderDimColor?: boolean;
   } = {},
 ): HistoryItemToolGroup {
   const toolCalls = Array.isArray(toolOrTools) ? toolOrTools : [toolOrTools];
-  const { borderTop, borderBottom } = options;
+  const { borderTop, borderBottom, borderColor, borderDimColor } = options;
 
   const toolDisplays = toolCalls.map((call): IndividualToolCallDisplay => {
     let description: string;
@@ -107,5 +109,7 @@ export function mapToDisplay(
     tools: toolDisplays,
     borderTop,
     borderBottom,
+    borderColor,
+    borderDimColor,
   };
 }
